# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Address:
    properties:
      city_id:
        type: string
      created_at:
        type: string
      detail:
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      postal_code:
        type: string
      province_id:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/domain.User'
        description: Relations
      user_id:
        type: integer
    required:
    - city_id
    - detail
    - name
    - phone
    - postal_code
    - province_id
    type: object
...
  domain.City:
    properties:
      id:
        type: string
      name:
        type: string
      province_id:
        type: string
    type: object
  domain.CreateAddressRequest:
    properties:
      city_id:
        type: string
      detail:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      postal_code:
        type: string
      province_id:
        type: string
    required:
    - city_id
    - detail
    - name
    - phone
    - postal_code
    - province_id
    type: object
...
  domain.Province:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  domain.UpdateAddressRequest:
    properties:
      city_id:
        type: string
      detail:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      postal_code:
        type: string
      province_id:
        type: string
    required:
    - city_id
    - detail
    - name
    - phone
    - postal_code
    - province_id
    type: object
...
  domain.User:
    properties:
      about:
        type: string
      city_id:
        type: integer
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      gender:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      job:
        type: string
      last_login_at:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      photo_url:
        type: string
      province_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    required:
    - email
    - name
    - phone
    type: object
  response.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_page:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: E-commerce API built with Go Fiber and Clean Architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Commerce API
  version: "1.0"
paths:
  /addresses:
    get:
      consumes:
      - application/json
      description: Get all addresses for the authenticated user with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Addresses retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Address'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get current user's addresses
      tags:
      - Addresses
    post:
      consumes:
      - application/json
      description: Create a new address for the authenticated user
      parameters:
      - description: Address creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateAddressRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Address created successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Address'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new address
      tags:
      - Addresses
  /addresses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing address (owner only)
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Address deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Address not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete an address
      tags:
      - Addresses
    get:
      consumes:
      - application/json
      description: Get a single address by its ID (owner only)
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Address retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Address'
              type: object
        "400":
          description: Invalid address ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Address not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get address by ID
      tags:
      - Addresses
    put:
      consumes:
      - application/json
      description: Update an existing address (owner only)
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: integer
      - description: Address update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Address updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Address'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Address not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update an address
      tags:
      - Addresses
...
  /regions/provinces:
    get:
      consumes:
      - application/json
      description: Get all provinces in Indonesia (public endpoint)
      produces:
      - application/json
      responses:
        "200":
          description: Provinces retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Province'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all provinces
      tags:
      - Regions
  /regions/provinces/{provinceId}/cities:
    get:
      consumes:
      - application/json
      description: Get all cities in a specific province (public endpoint)
      parameters:
      - description: Province ID
        in: path
        name: provinceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cities retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.City'
                  type: array
              type: object
        "400":
          description: Province ID is required
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get cities by province
      tags:
      - Regions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"